chore : upgrade dependencies to latest stable
Bump every dependency to its latest stable release, verified against
maven-metadata.xml (the search.maven.org solr index is stale):

- Gradle 9.1.0 -> 9.5.1
- Kotlin 2.3.21 -> 2.4.0
- Spring Boot 4.0.6 -> 4.1.0
- Spring Cloud 2025.1.1 -> 2025.1.2 (first patch supporting Boot 4.1)
- Kotest 6.1.11 -> 6.2.0, mockk 1.14.9 -> 1.14.11
- Redisson 3.52.0 -> 4.6.0 (major), reactor-kotlin-extensions 1.3.0 -> 1.3.1
- kotlin-logging 7.0.14 -> 8.0.4 (major)

Code migration for the major bumps:
- RScript.ReturnType.INTEGER -> LONG (Redisson 4.0 enum rename)
- move password/keepAlive/tcpNoDelay from per-server to top-level Config
  (per-server setters deprecated in Redisson 4.x)
- exclude RedissonAutoConfigurationV4 as well as V2
- audienceValidator JwtClaimValidator<List<String>?> -> <List<String>>
  for Spring Security 7's JSpecify non-null bound
- drop redundant -Xannotation-default-target compiler arg (default in Kotlin 2.4)

clean build green (tests pass, zero warnings); bootJar starts with health UP.

+++ src/main/kotlin/dev/notypie/gateway/modules/redis/ReactiveRedissonClientModule.kt
@@ -77,7 +77,7 @@ class ReactiveRedissonClientModule(
                     end
                     return v
                     """.trimIndent(),
-                    RScript.ReturnType.INTEGER,
+                    RScript.ReturnType.LONG,
                     listOf(key),
                     count.toString(),
                     ttlSeconds.toString(),
